src/third_party/blink/web_tests/http/tests/cachestorage/padding.html

<!doctype html>
<title>Cache Storage: verify cache_storage padding behavior</title>
<link rel="help" href="https://w3c.github.io/ServiceWorker/#cache-interface">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

async function usage(cache, url, init, clone) {
  const i = { mode: 'no-cors' };
  let r = await fetch(url, {
    mode: 'no-cors',
    credentials: init.credentials,
    method: init.method,
  });
  assert_equals('opaque', r.type);
  if (clone)
    r = r.clone();
  // Note, this stores the response under a Request key that has different
  // attributes than what we fetched above.  This is purposeful to ensure
  // that we are varying on how the Response was loaded and not on the Request
  // key.
  await cache.put(url, r);
  return (await navigator.storage.estimate()).usage;
}

promise_test(async t => {
  const cache = await caches.open('padding');
  const url = 'http://localhost:8000/cachestorage/resources/simple.txt';

  // Get the reported disk usage for an opaque response without credentials.
  const usage1 = await usage(cache, url, { credentials: 'omit' });
  const usage2 = await usage(cache, url, { credentials: 'same-origin' });

  // Get the reported disk usage for an opaque response with the same URL,
  // but also with credentials enabled.
  const usage3 = await usage(cache, url, { credentials: 'include' });

  assert_equals(usage1, usage2,
                "Credentials mode 'omit' and 'same-origin' should have the " +
                "same padding size since credentials are not sent.");
  assert_not_equals(usage1, usage3,
                    "Credentials mode 'include' should result in a different " +
                    "padding size.");
}, 'Cache padding varies based on if Response was loaded with credentials.');

promise_test(async t => {
  const cache = await caches.open('padding');
  const url = 'http://localhost:8000/cachestorage/resources/simple.txt';

  const usage1 = await usage(cache, url, { method: 'GET' });
  const usage1b = await usage(cache, url, { method: 'GET' });
  assert_equals(usage1, usage1b,
                "Usage should be same if the request does not vary.");

  const usage2 = await usage(cache, url, { method: 'HEAD' });
  const usage2b = await usage(cache, url, { method: 'HEAD' });
  assert_equals(usage2, usage2b,
                "Usage should be same if the request does not vary.");

  const usage3 = await usage(cache, url, { method: 'POST' });
  const usage3b = await usage(cache, url, { method: 'POST' });
  assert_equals(usage3, usage3b,
                "Usage should be same if the request does not vary.");

  // It should only be possible to create opaque responses with CORS
  // safelisted methods.
  await promise_rejects_js(t, TypeError, usage(cache, url, { method: 'OPTIONS' }));
  await promise_rejects_js(t, TypeError, usage(cache, url, { method: 'CONNECT' }));
  await promise_rejects_js(t, TypeError, usage(cache, url, { method: 'TRACE' }));
  await promise_rejects_js(t, TypeError, usage(cache, url, { method: 'TRACK' }));
  await promise_rejects_js(t, TypeError, usage(cache, url, { method: 'PUT' }));
  await promise_rejects_js(t, TypeError, usage(cache, url, { method: 'PATCH' }));
  await promise_rejects_js(t, TypeError, usage(cache, url, { method: 'FOO' }));

  assert_not_equals(usage1, usage2,
                    "Methods 'GET' and 'HEAD' should result in different " +
                    "padding sizes.");
  assert_not_equals(usage1, usage3,
                    "Methods 'GET' and 'POST' should result in different " +
                    "padding sizes.");
  assert_not_equals(usage2, usage3,
                    "Methods 'HEAD' and 'POST' should result in different " +
                    "padding sizes.");
}, 'Cache padding varies based on Response original request method.');

promise_test(async t => {
  const cache = await caches.open('padding');
  const url = 'http://localhost:8000/cachestorage/resources/simple.txt';
  const init = { method: 'GET', credentials: 'include' };

  const usage1 = await usage(cache, url, init, /* clone=*/ false);
  const usage2 = await usage(cache, url, init, /* clone=*/ true);
  assert_equals(usage1, usage2, "Usage should not vary based on cloning.");
}, 'Cache padding does not change when cloned.');

</script>
